signals.py and receivers.py
for signal definitions and handler logic
