+++ foodsaving/history/receivers.py
@@ -3,9 +3,8 @@ from django.dispatch import receiver
 from foodsaving.history.models import History, HistoryTypus
 from foodsaving.groups.signals import post_group_join, pre_group_leave, post_group_modify, post_group_create
 from foodsaving.stores.api import post_store_delete, pre_pickup_delete, pre_series_delete
-from foodsaving.stores.serializers import post_store_create, post_store_modify, post_pickup_create, \
-    post_pickup_modify, post_series_create, post_series_modify, post_pickup_join, post_pickup_leave
-from foodsaving.stores.models import pickup_done, pickup_missed
+from foodsaving.stores.signals import post_pickup_create, post_pickup_modify, post_pickup_join, post_pickup_leave, \
+    post_series_create, post_series_modify, post_store_create, post_store_modify, pickup_done, pickup_missed
 
 
 def make_handler(typus):
